pyproject.toml
[tool.poetry]
name = "django-federation-auditlog"
version = "0.2.2"
description = ""
authors = ["Anderson Marques <anderson89marques@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "==3.11.1"
djangorestframework-simplejwt = "==5.0.0"
django-auditlog = "==1.0.0"

[tool.poetry.dev-dependencies]
black = "^22.8b1"
flake8 = "^3.8.3"


[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
